bin/requirements.txt
discord.py
dotenv
pyautogui
signal
subprocess